+++ src/Application/GuestAccess/GuestSessionBootstrap.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EventFlow\Application\GuestAccess;
+
+interface GuestSessionBootstrap
+{
+    public function bootstrap(string $rawCredential, GuestCredentialType $type): GuestSessionCredentials;
+}